fix modification count in table
